Able to now determine whether a post is new or not.
This is achived by checking the updated or published dates. If neither exist
then it falls back and checks if it has previously seem the post's link.

This system currently does not have a method of handling if the Feed's items
do not either the dates or a link. More fallbacks may be required in the future.